+++ crates/holochain_p2p/src/spawn.rs
@@ -238,8 +238,8 @@ impl Default for HolochainP2pConfig {
     }
 }
 
-/// See [NetworkCompatParams::proto_ver].
-pub const HCP2P_PROTO_VER: u32 = 2;
+/// See [`NetworkCompatParams::proto_ver`].
+pub const HCP2P_PROTO_VER: u32 = 3;
 
 /// Some parameters used as part of a protocol compatibility check during tx5 preflight
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, serde::Serialize, serde::Deserialize)]